requirements/local.txt
-r base.txt
ipython
django-debug-toolbar
pytest-django
factory_boy
